+++ statustracker-server/src/utils.rs
@@ -14,14 +14,14 @@ pub fn get_hour_timestamp(t: SystemTime) -> HourTimestamp {
 #[derive(Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq, Default)]
 pub struct BitField64(pub i32, pub i32);
 impl BitField64 {
-    pub fn turn_on(&mut self, i: usize) {
+    pub fn turn_on(&mut self, i: i32) {
         if i < 30 {
             self.0 |= 1 << i
         } else {
             self.1 |= 1 << (i - 30)
         }
     }
-    pub fn is_on(self, i: usize) -> bool {
+    pub fn is_on(self, i: i32) -> bool {
         if i < 30 {
             self.0 & (1 << i) != 0
         } else {
